Fixed deleting issues, but slow again
The issue that was causing deletes to not be performed was that since I
was bucketing, it wouldn't recieve that update. This means that I have
to be okay with the buffering issues. I believe the next step to trying
to solve this issue is switching the server to be a UDP based server,
rather than a websocket.