Finished new server setup
The new UDP server was having issues that would cause only one bounding
box to be updated at at time. The issue was that the server was popping
all the messages off the queue and only saving the last one, so I
updated it to process all of the queue. Additionally, I switched the
timer back to being based on bounding box updates instead of actual
time. The new server architecture is now complete.